+++ packages/datasource/src/infra/index.ts
@@ -1,4 +1,4 @@
 export * from "./config";
 export * from "./postgres/postgres.adapter";
-export * from "./postgres/orders-mutation.adapter";
-export * from "./postgres/org-scope.adapter";
+export * from "./postgres/postgres-demo-orders-mutation.adapter";
+export * from "./postgres/postgres-org-scope.adapter";